added search functionality on petition title
made some fixes to variable types

+++ src/main/java/com/example/johnspetitions/controller/SignatureController.java
@@ -26,7 +26,7 @@ public class SignatureController {
 
     @PostMapping("/sign")
     @ResponseBody
-    public ResponseEntity<String> signPetition(@RequestParam Long petitionId,
+    public ResponseEntity<String> signPetition(@RequestParam Integer petitionId,
                                                @RequestParam String signerName,
                                                @RequestParam String signerEmail) {
         try {
